+++ src/App.tsx
@@ -1,10 +1,17 @@
 import React from 'react';
+import { TodosProvider } from './components/TodosContext';
 import { TodoApp } from './components/TodoApp';
+import { Main } from './components/Main';
+import { Footer } from './components/Footer';
 
 export const App: React.FC = () => {
   return (
     <div className="todoapp">
-      <TodoApp />
+      <TodosProvider>
+        <TodoApp />
+        <Main />
+        <Footer />
+      </TodosProvider>
     </div>
   );
 };
